tasks/cartographer/app/requirements.txt
flask
openlocationcode
piexif
scipy
pillow
haversine
requests
